Remove type_declare.h
