+++ src/kv/generic_serialise_wrapper.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "ds/buffer.h"
-#include "type_declare.h"
+#include "kv_types.h"
 
 #include <optional>
