+++ src/node/scripts.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "script.h"
 #include "kv/map.h"
+#include "script.h"
 
 namespace ccf
 {
